+++ docs/release_notes.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 * __Template类中新增layout参数，以支持排版布局的灵活配置。__
+* __Template类移除font_size参数，字体大小改为使用font参数的size属性表示。__
 
 ## v6.0.0 (2020-10-18)
 * 提供对Python3.9的支持